+++ EngineCode/Engine/render_DX9_frame_pipeline.cpp
@@ -14,8 +14,8 @@
 #include "user_interface.h"
 #include "main_window.h"
 ///////////////////////////////////////////////////////////////
-extern UINT g_ScreenWidth;
-extern UINT g_ScreenHeight;
+extern uint16_t g_ScreenWidth;
+extern uint16_t g_ScreenHeight;
 ///////////////////////////////////////////////////////////////
 void CRenderDX9::OnFrame()
 {
